+++ src/components/breadcrumbItem.js
@@ -66,7 +66,7 @@
       breadcrumbItem
     );
   })(),
-  styles: B => t => {
+  styles: (B) => (t) => {
     const { Styling } = B;
     const style = new Styling(t);
     return {