+++ src/components/dataTable.js
@@ -12,6 +12,7 @@
       ModelProvider,
       useAllQuery,
       useFilter,
+      useRelation,
       useText,
       Icon,
     } = B;
@@ -126,11 +127,11 @@
     const [stylesProps, setStylesProps] = useState(null);
 
     const deepMerge = (...objects) => {
-      const isObject = item =>
+      const isObject = (item) =>
         item && typeof item === 'object' && !Array.isArray(item);
 
       return objects.reduce((accumulator, object) => {
-        Object.keys(object).forEach(key => {
+        Object.keys(object).forEach((key) => {
           const accumulatorValue = accumulator[key];
           const value = object[key];
 
@@ -151,7 +152,7 @@
       path = [searchProperty.id].flat();
     }
 
-    const transformValue = value => {
+    const transformValue = (value) => {
       if (value instanceof Date) {
         return value.toISOString();
       }
@@ -181,7 +182,7 @@
 
     let interactionFilters = {};
 
-    const isEmptyValue = value =>
+    const isEmptyValue = (value) =>
       !value || (Array.isArray(value) && value.length === 0);
 
     const clauses = Object.entries(interactionFilter)
@@ -192,7 +193,7 @@
           if (isLast) {
             return Array.isArray(value)
               ? {
-                  _or: value.map(el => ({
+                  _or: value.map((el) => ({
                     [field]: { [property.operator]: el },
                   })),
                 }
@@ -226,7 +227,12 @@
     const where = useFilter(completeFilter);
 
     // TODO: move model to skip
-    const { loading, error, data, refetch } = useAllQuery(
+    const {
+      loading: queryLoading,
+      error,
+      data: queryData,
+      refetch,
+    } = useAllQuery(
       model,
       {
         rawFilter: where,
@@ -250,6 +256,14 @@
       !model,
     );
 
+    const { hasResults, data: relationData } = useRelation(
+      model,
+      {},
+      typeof model === 'string' || !model,
+    );
+    const data = hasResults ? relationData : queryData;
+    const loading = hasResults ? false : queryLoading;
+
     useEffect(() => {
       if (!isDev && data) {
         if (pagination !== 'never') {
@@ -270,7 +284,7 @@
           ) {
             setResults(data.results);
           } else {
-            setResults(prev => [...prev, ...data.results]);
+            setResults((prev) => [...prev, ...data.results]);
           }
           fetchingNextSet.current = false;
           setNewSearch(false);
@@ -310,7 +324,7 @@
       }
     });
 
-    B.defineFunction('SetSearchValue', event => {
+    B.defineFunction('SetSearchValue', (event) => {
       setSearch(event.target.value);
     });
 
@@ -331,7 +345,7 @@
           const childrenLenght = children.length;
           const collection = Array.from(repeaterRef.current.children);
           collection
-            .filter(item => item.tagName === 'DIV')
+            .filter((item) => item.tagName === 'DIV')
             .forEach((item, index) => {
               if (
                 ((Math.ceil((index + 1) / childrenLenght) * childrenLenght) /
@@ -382,7 +396,7 @@
       setPage(newPage);
     };
 
-    const handleChangeRowsPerPage = event => {
+    const handleChangeRowsPerPage = (event) => {
       if (loading || error) return;
       setRowsPerPage(parseInt(event.target.value, 10));
       setPage(0);
@@ -398,7 +412,7 @@
       });
     };
 
-    const handleSearch = event => {
+    const handleSearch = (event) => {
       setSearch(event.target.value);
     };
 
@@ -413,7 +427,7 @@
 
     const renderTableHead = () => {
       if ((loading && !loadOnScroll) || error) {
-        return Array.from(Array(children.length).keys()).map(colIdx => (
+        return Array.from(Array(children.length).keys()).map((colIdx) => (
           <TableCell key={colIdx}>
             <div className={classes.skeleton}>
               {error && displayError && error.message}
@@ -430,9 +444,9 @@
 
     const tableContentModel = () => {
       if ((loading && !loadOnScroll) || error) {
-        return Array.from(Array(rowsPerPage).keys()).map(idx => (
+        return Array.from(Array(rowsPerPage).keys()).map((idx) => (
           <TableRow key={idx} classes={{ root: classes.bodyRow }}>
-            {Array.from(Array(children.length).keys()).map(colIdx => (
+            {Array.from(Array(children.length).keys()).map((colIdx) => (
               <TableCell key={colIdx}>
                 <div className={classes.skeleton} />
               </TableCell>
@@ -441,10 +455,10 @@
         ));
       }
 
-      return results.map(value => (
-        <ModelProvider value={value} id={model}>
+      return results.map((value) => (
+        <ModelProvider key={model} value={value} id={model}>
           <InteractionScope model={model}>
-            {context => (
+            {(context) => (
               <TableRow
                 key={value[0]}
                 classes={{ root: classes.bodyRow }}
@@ -475,7 +489,7 @@
         return tableContentModel();
       }
 
-      return Array.from(Array(amountOfRows).keys()).map(idx => (
+      return Array.from(Array(amountOfRows).keys()).map((idx) => (
         <TableRow key={idx} classes={{ root: classes.bodyRow }}>
           {children}
         </TableRow>
@@ -487,7 +501,7 @@
         const fetchNextSet = () => {
           fetchingNextSet.current = true;
           if (!initialRender.current) {
-            setSkip(prev => prev + autoLoadTakeAmountNum);
+            setSkip((prev) => prev + autoLoadTakeAmountNum);
           }
           initialRender.current = false;
         };
@@ -500,10 +514,10 @@
             tableContainerElement.scrollHeight <= parent.clientHeight &&
             initialTimesFetched < 5
           ) {
-            setInitialTimesFetched(prev => prev + 1);
+            setInitialTimesFetched((prev) => prev + 1);
             fetchNextSet();
           }
-          const scrollEvent = e => {
+          const scrollEvent = (e) => {
             const { scrollTop, clientHeight, scrollHeight } = e.target;
             const offset = scrollHeight / 5;
             const hitBottom = scrollTop + clientHeight >= scrollHeight - offset;
@@ -562,20 +576,22 @@
               setShowPagination(true);
             }
             break;
-          default:
           case 'always':
+          default:
             setShowPagination(true);
             break;
         }
       }
     }, [data, rowsPerPage]);
 
+    const isRelation = !isDev && typeof model !== 'string';
+
     useEffect(() => {
       let amount = 0;
       if (hasToolbar) {
         amount += toolbarRef.current.clientHeight;
       }
-      if (showPagination) {
+      if (showPagination && !isRelation) {
         amount += paginationRef.current.clientHeight;
       }
       let style;
@@ -643,7 +659,7 @@
               )}
             </Table>
           </TableContainer>
-          {showPagination && (
+          {showPagination && !isRelation && (
             <TablePagination
               ref={paginationRef}
               classes={{ root: classes.pagination }}
@@ -664,7 +680,7 @@
       </div>
     );
   })(),
-  styles: B => theme => {
+  styles: (B) => (theme) => {
     const { env, mediaMinWidth, Styling } = B;
     const style = new Styling(theme);
     const isDev = env === 'dev';