+++ src/components/drawerBar.js
@@ -110,7 +110,7 @@
       </div>
     );
   })(),
-  styles: B => t => {
+  styles: (B) => (t) => {
     const { env, mediaMinWidth, Styling } = B;
     const style = new Styling(t);
     const getSpacing = (idx, device = 'Mobile') =>