+++ src/components/radio.js
@@ -33,6 +33,7 @@
       getCustomModelAttribute,
       getProperty,
       useAllQuery,
+      useRelation,
       useText,
     } = B;
     const isDev = env === 'dev';
@@ -66,7 +67,7 @@
       : Number(componentValue);
 
     // maintain the type of the value
-    const getValue = val => (isNaN(Number(val)) ? val : Number(val));
+    const getValue = (val) => (isNaN(Number(val)) ? val : Number(val));
     const [value, setValue] = useState(getValue(componentValue));
     const [errorState, setErrorState] = useState(false);
     const [afterFirstInvalidation, setAfterFirstInvalidation] = useState(false);
@@ -99,7 +100,12 @@
           }, {})
         : {};
 
-    const { loading, error: err, data, refetch } = useAllQuery(
+    const {
+      loading: queryLoading,
+      error: err,
+      data: queryData,
+      refetch,
+    } = useAllQuery(
       model,
       {
         filter,
@@ -123,6 +129,15 @@
       },
       !model,
     );
+
+    const { hasResults, data: relationData } = useRelation(
+      model,
+      {},
+      typeof model === 'string' || !model,
+    );
+
+    const data = hasResults ? relationData : queryData;
+    const loading = hasResults ? false : queryLoading;
 
     useEffect(() => {
       if (mounted.current) {
@@ -165,15 +180,15 @@
         return listValues.map(({ value: v }) => renderRadio(v, v));
       }
       if (optionType === 'static') {
-        radioValues = radioData.map(option => option);
-        return radioData.map(option => renderRadio(getValue(option), option));
+        radioValues = radioData.map((option) => option);
+        return radioData.map((option) => renderRadio(getValue(option), option));
       }
       if (isDev) return renderRadio('value', 'Placeholder');
       if (loading) return <span>Loading...</span>;
       if (err && displayError) return <span>{err.message}</span>;
 
-      radioValues = results.map(item => item[valueProperty.name]);
-      return results.map(item =>
+      radioValues = results.map((item) => item[valueProperty.name]);
+      return results.map((item) =>
         renderRadio(item[valueProperty.name], item[labelProperty.name]),
       );
     };
@@ -187,7 +202,7 @@
       setHelper(message);
     };
 
-    const handleChange = evt => {
+    const handleChange = (evt) => {
       if (afterFirstInvalidation) {
         handleValidation();
       }
@@ -247,7 +262,7 @@
       FormControl
     );
   })(),
-  styles: B => t => {
+  styles: (B) => (t) => {
     const { color: colorFunc, Styling } = B;
     const style = new Styling(t);
     const getOpacColor = (col, val) => colorFunc.alpha(col, val);