+++ src/components/select.js
@@ -33,6 +33,7 @@
       getCustomModelAttribute,
       getProperty,
       useAllQuery,
+      useRelation,
       useText,
     } = B;
     const { TextField, MenuItem } = window.MaterialUI.Core;
@@ -76,7 +77,7 @@
     const validationMessageText = useText(validationValueMissing);
     const dataComponentAttributeValue = useText(dataComponentAttribute);
 
-    const transformValue = arg => {
+    const transformValue = (arg) => {
       if (arg instanceof Date) {
         return arg.toISOString();
       }
@@ -85,11 +86,11 @@
     };
 
     const deepMerge = (...objects) => {
-      const isObject = item =>
+      const isObject = (item) =>
         item && typeof item === 'object' && !Array.isArray(item);
 
       return objects.reduce((accumulator, object) => {
-        Object.keys(object).forEach(key => {
+        Object.keys(object).forEach((key) => {
           const accumulatorValue = accumulator[key];
           const valueArg = object[key];
 
@@ -120,7 +121,7 @@
 
     let interactionFilters = {};
 
-    const isEmptyValue = arg =>
+    const isEmptyValue = (arg) =>
       !arg || (Array.isArray(arg) && arg.length === 0);
 
     const clauses = Object.entries(interactionFilter)
@@ -131,7 +132,7 @@
           if (isLast) {
             return Array.isArray(valueArg)
               ? {
-                  _or: valueArg.map(el => ({
+                  _or: valueArg.map((el) => ({
                     [field]: { [propertyArg.operator]: el },
                   })),
                 }
@@ -147,7 +148,12 @@
 
     const completeFilter = deepMerge(filter, interactionFilters);
 
-    const { loading, error, data, refetch } = useAllQuery(
+    const {
+      loading: queryLoading,
+      error,
+      data: queryData,
+      refetch,
+    } = useAllQuery(
       model,
       {
         filter: completeFilter,
@@ -171,6 +177,15 @@
       },
       !model,
     );
+
+    const { hasResults, data: relationData } = useRelation(
+      model,
+      {},
+      typeof model === 'string' || !model,
+    );
+
+    const data = hasResults ? relationData : queryData;
+    const loading = hasResults ? false : queryLoading;
 
     useEffect(() => {
       if (mounted.current) {
@@ -224,7 +239,7 @@
       setHelper(message);
     };
 
-    const handleChange = event => {
+    const handleChange = (event) => {
       const {
         target: { value: eventValue },
       } = event;
@@ -257,7 +272,7 @@
         ));
       }
       if (optionType === 'static') {
-        return selectOptions.split('\n').map(option => (
+        return selectOptions.split('\n').map((option) => (
           <MenuItem key={option} value={option}>
             {option}
           </MenuItem>
@@ -266,7 +281,7 @@
       if (loading) return <span>Loading...</span>;
       if (error && displayError) return <span>{error.message}</span>;
       return (results || []).map(
-        item =>
+        (item) =>
           propName &&
           labelName && (
             <MenuItem key={item.id} value={item[propName]}>
@@ -316,7 +331,7 @@
 
     return isDev ? <div className={classes.root}>{SelectCmp}</div> : SelectCmp;
   })(),
-  styles: B => t => {
+  styles: (B) => (t) => {
     const { Styling } = B;
     const style = new Styling(t);
     return {
@@ -382,20 +397,22 @@
             '!important',
           ],
           '&:hover': {
-            '& .MuiOutlinedInput-notchedOutline, & .MuiFilledInput-underline, & .MuiInput-underline': {
-              borderColor: ({ options: { borderHoverColor } }) => [
-                style.getColor(borderHoverColor),
-                '!important',
-              ],
-            },
+            '& .MuiOutlinedInput-notchedOutline, & .MuiFilledInput-underline, & .MuiInput-underline':
+              {
+                borderColor: ({ options: { borderHoverColor } }) => [
+                  style.getColor(borderHoverColor),
+                  '!important',
+                ],
+              },
           },
           '&.Mui-focused, &.Mui-focused:hover': {
-            '& .MuiOutlinedInput-notchedOutline, & .MuiFilledInput-underline, & .MuiInput-underline': {
-              borderColor: ({ options: { borderFocusColor } }) => [
-                style.getColor(borderFocusColor),
-                '!important',
-              ],
-            },
+            '& .MuiOutlinedInput-notchedOutline, & .MuiFilledInput-underline, & .MuiInput-underline':
+              {
+                borderColor: ({ options: { borderFocusColor } }) => [
+                  style.getColor(borderFocusColor),
+                  '!important',
+                ],
+              },
           },
           '& fieldset': {
             top: ({ options: { hideLabel } }) => (hideLabel ? 0 : null),
@@ -415,12 +432,13 @@
             '!important',
           ],
         },
-        '& .MuiOutlinedInput-notchedOutline, & .MuiFilledInput-underline, & .MuiInput-underline': {
-          borderColor: ({ options: { borderColor } }) => [
-            style.getColor(borderColor),
-            '!important',
-          ],
-        },
+        '& .MuiOutlinedInput-notchedOutline, & .MuiFilledInput-underline, & .MuiInput-underline':
+          {
+            borderColor: ({ options: { borderColor } }) => [
+              style.getColor(borderColor),
+              '!important',
+            ],
+          },
         '& .MuiInput-underline, & .MuiFilledInput-underline': {
           '&::before, &::after': {
             borderColor: ({ options: { borderColor } }) => [
@@ -436,43 +454,47 @@
               ],
             },
           },
-          '&.Mui-focused::before, &.Mui-focused::after, &.Mui-focused:hover::before, &.Mui-focused:hover::after': {
-            borderColor: ({ options: { borderFocusColor } }) => [
-              style.getColor(borderFocusColor),
-              '!important',
-            ],
-          },
-        },
-        '& .MuiInputBase-root.Mui-error, & .MuiInputBase-root.Mui-error:hover, & .MuiInputBase-root.Mui-error.Mui-focused, & .MuiInputBase-root.Mui-error.Mui-focused:hover': {
-          '& .MuiOutlinedInput-notchedOutline, & .MuiFilledInput-underline, & .MuiInput-underline': {
-            borderColor: ({ options: { errorColor } }) => [
-              style.getColor(errorColor),
-              '!important',
-            ],
-          },
-          '&.MuiInput-underline, &.MuiFilledInput-underline': {
-            '&::before, &::after': {
-              borderColor: ({ options: { errorColor } }) => [
-                style.getColor(errorColor),
+          '&.Mui-focused::before, &.Mui-focused::after, &.Mui-focused:hover::before, &.Mui-focused:hover::after':
+            {
+              borderColor: ({ options: { borderFocusColor } }) => [
+                style.getColor(borderFocusColor),
                 '!important',
               ],
             },
-            '&:hover': {
+        },
+        '& .MuiInputBase-root.Mui-error, & .MuiInputBase-root.Mui-error:hover, & .MuiInputBase-root.Mui-error.Mui-focused, & .MuiInputBase-root.Mui-error.Mui-focused:hover':
+          {
+            '& .MuiOutlinedInput-notchedOutline, & .MuiFilledInput-underline, & .MuiInput-underline':
+              {
+                borderColor: ({ options: { errorColor } }) => [
+                  style.getColor(errorColor),
+                  '!important',
+                ],
+              },
+            '&.MuiInput-underline, &.MuiFilledInput-underline': {
               '&::before, &::after': {
                 borderColor: ({ options: { errorColor } }) => [
                   style.getColor(errorColor),
                   '!important',
                 ],
               },
-            },
-            '&.Mui-focused::before, &.Mui-focused::after, &.Mui-focused:hover::before, &.Mui-focused:hover::after': {
-              borderColor: ({ options: { errorColor } }) => [
-                style.getColor(errorColor),
-                '!important',
-              ],
+              '&:hover': {
+                '&::before, &::after': {
+                  borderColor: ({ options: { errorColor } }) => [
+                    style.getColor(errorColor),
+                    '!important',
+                  ],
+                },
+              },
+              '&.Mui-focused::before, &.Mui-focused::after, &.Mui-focused:hover::before, &.Mui-focused:hover::after':
+                {
+                  borderColor: ({ options: { errorColor } }) => [
+                    style.getColor(errorColor),
+                    '!important',
+                  ],
+                },
             },
           },
-        },
       },
     };
   },