+++ src/components/slider.js
@@ -4,12 +4,8 @@
   allowedTypes: [],
   orientation: 'HORIZONTAL',
   jsx: (() => {
-    const {
-      Slider,
-      InputLabel,
-      FormHelperText,
-      FormControl,
-    } = window.MaterialUI.Core;
+    const { Slider, InputLabel, FormHelperText, FormControl } =
+      window.MaterialUI.Core;
     const { env, getCustomModelAttribute, useText } = B;
     const isDev = env === 'dev';
     const {
@@ -32,6 +28,7 @@
       value: defaultValue = [],
     } = customModelAttributeObj;
     const [currentValue, setCurrentValue] = useState(useText(defaultValue));
+    const [interactionFilter, setInteractionFilter] = useState({});
     const [errorState, setErrorState] = useState(error);
     const [helper, setHelper] = useState(useText(helperText));
     const [afterFirstInvalidation, setAfterFirstInvalidation] = useState(false);
@@ -44,6 +41,7 @@
     const dataComponentAttributeValue = useText(dataComponentAttribute);
     const helperTextResolved = useText(helperText);
     const validationValueMissingText = useText(validationValueMissing);
+    const mounted = useRef(false);
 
     const handleValidation = () => {
       const hasError = required && !currentValue;
@@ -60,10 +58,17 @@
       handleValidation();
     };
 
-    const setValue = (_event, value) => {
-      setCurrentValue(value);
-      if (afterFirstInvalidation) {
-        handleValidation();
+    const changeHandler = (_event, value) => {
+      if (!isDev) {
+        if (afterFirstInvalidation) {
+          handleValidation();
+        }
+        setCurrentValue(value);
+      }
+    };
+    const changeCommitHandler = () => {
+      if (mounted.current && !isDev) {
+        B.triggerEvent('onChange', currentValue);
       }
     };
 
@@ -77,6 +82,38 @@
         setHelper(helperTextResolved);
       }
     }, [isDev, helperTextResolved]);
+    useEffect(() => {
+      mounted.current = true;
+      return () => {
+        if (!isDev) {
+          mounted.current = false;
+        }
+      };
+    }, []);
+
+    /**
+     * @name Filter
+     * @param {Property} property
+     * @returns {Void}
+     */
+    B.defineFunction(
+      'Filter',
+      ({ event, property: propertyArg, interactionId }) => {
+        setInteractionFilter({
+          ...interactionFilter,
+          [interactionId]: {
+            property: propertyArg,
+            value: event.target ? event.target.value : event,
+          },
+        });
+      },
+    );
+
+    B.defineFunction('ResetFilter', () => {
+      setInteractionFilter({});
+    });
+
+    B.defineFunction('Reset', () => setCurrentValue(useText(defaultValue)));
 
     const sliderInput = (
       <div className={classes.root}>
@@ -100,7 +137,8 @@
               min={startNumber}
               max={endNumber}
               marks={stepNumber ? marks : false}
-              onChange={setValue}
+              onChange={changeHandler}
+              onChangeCommitted={changeCommitHandler}
               disabled={isDev || disable}
               classes={{
                 root: classes.sliderRoot,
@@ -133,7 +171,7 @@
       <div className={classes.dev}>{sliderInput}</div>
     );
   })(),
-  styles: B => t => {
+  styles: (B) => (t) => {
     const { mediaMinWidth, Styling } = B;
     const style = new Styling(t);
     const getSpacing = (idx, device = 'Mobile') =>