+++ src/components/textField.js
@@ -7,6 +7,7 @@
     const {
       autoComplete,
       disabled,
+      spellCheck,
       error,
       multiline,
       rows,
@@ -99,7 +100,7 @@
     const defaultValueText = useText(defaultValue);
     const dataComponentAttributeValue = useText(dataComponentAttribute);
 
-    const validationMessage = validityObject => {
+    const validationMessage = (validityObject) => {
       if (validityObject.customError && patternMismatchMessage) {
         return patternMismatchMessage;
       }
@@ -130,19 +131,19 @@
       return '';
     };
 
-    const handleValidation = validation => {
+    const handleValidation = (validation) => {
       setErrorState(!validation.valid);
       const message = validationMessage(validation) || helperTextResolved;
       setHelper(message);
     };
 
-    const onKeyDown = event => {
+    const onKeyDown = (event) => {
       if (isNumberType && (event.key === '.' || event.key === ',')) {
         event.preventDefault();
       }
     };
 
-    const customPatternValidation = target => {
+    const customPatternValidation = (target) => {
       const { value: eventValue, validity } = target;
       if (!pattern) {
         return validity;
@@ -157,7 +158,7 @@
       };
     };
 
-    const changeHandler = event => {
+    const changeHandler = (event) => {
       const { target } = event;
       let { validity: validation } = target;
       const { value: eventValue } = target;
@@ -176,7 +177,7 @@
       B.triggerEvent('onChange', value);
     };
 
-    const blurHandler = event => {
+    const blurHandler = (event) => {
       const { target } = event;
       let { validity: validation } = target;
 
@@ -188,7 +189,7 @@
       handleValidation(validation);
     };
 
-    const invalidHandler = event => {
+    const invalidHandler = (event) => {
       event.preventDefault();
       const {
         target: {
@@ -209,7 +210,7 @@
       togglePassword(!showPassword);
     };
 
-    const handleMouseDownPassword = event => {
+    const handleMouseDownPassword = (event) => {
       event.preventDefault();
     };
 
@@ -304,6 +305,7 @@
             min: validMinvalue,
             max: validMaxvalue,
             tabIndex: isDev ? -1 : undefined,
+            spellCheck,
           }}
           data-component={dataComponentAttributeValue}
         />
@@ -321,7 +323,7 @@
       TextFieldCmp
     );
   })(),
-  styles: B => t => {
+  styles: (B) => (t) => {
     const { Styling } = B;
     const style = new Styling(t);
     return {
@@ -378,20 +380,22 @@
               '!important',
             ],
           '&:hover': {
-            '& .MuiOutlinedInput-notchedOutline, & .MuiFilledInput-underline, & .MuiInput-underline': {
-              borderColor: ({ options: { borderHoverColor } }) => [
-                style.getColor(borderHoverColor),
-                '!important',
-              ],
-            },
+            '& .MuiOutlinedInput-notchedOutline, & .MuiFilledInput-underline, & .MuiInput-underline':
+              {
+                borderColor: ({ options: { borderHoverColor } }) => [
+                  style.getColor(borderHoverColor),
+                  '!important',
+                ],
+              },
           },
           '&.Mui-focused, &.Mui-focused:hover': {
-            '& .MuiOutlinedInput-notchedOutline, & .MuiFilledInput-underline, & .MuiInput-underline': {
-              borderColor: ({ options: { borderFocusColor } }) => [
-                style.getColor(borderFocusColor),
-                '!important',
-              ],
-            },
+            '& .MuiOutlinedInput-notchedOutline, & .MuiFilledInput-underline, & .MuiInput-underline':
+              {
+                borderColor: ({ options: { borderFocusColor } }) => [
+                  style.getColor(borderFocusColor),
+                  '!important',
+                ],
+              },
           },
           '& fieldset': {
             top: ({ options: { hideLabel } }) => (hideLabel ? 0 : null),
@@ -420,12 +424,13 @@
             '!important',
           ],
         },
-        '& .MuiOutlinedInput-notchedOutline, & .MuiFilledInput-underline, & .MuiInput-underline': {
-          borderColor: ({ options: { borderColor } }) => [
-            style.getColor(borderColor),
-            '!important',
-          ],
-        },
+        '& .MuiOutlinedInput-notchedOutline, & .MuiFilledInput-underline, & .MuiInput-underline':
+          {
+            borderColor: ({ options: { borderColor } }) => [
+              style.getColor(borderColor),
+              '!important',
+            ],
+          },
         '& .MuiInput-underline, & .MuiFilledInput-underline': {
           '&::before, &::after': {
             borderColor: ({ options: { borderColor } }) => [
@@ -441,43 +446,47 @@
               ],
             },
           },
-          '&.Mui-focused::before, &.Mui-focused::after, &.Mui-focused:hover::before, &.Mui-focused:hover::after': {
-            borderColor: ({ options: { borderFocusColor } }) => [
-              style.getColor(borderFocusColor),
-              '!important',
-            ],
-          },
-        },
-        '& .MuiInputBase-root.Mui-error, & .MuiInputBase-root.Mui-error:hover, & .MuiInputBase-root.Mui-error.Mui-focused, & .MuiInputBase-root.Mui-error.Mui-focused:hover': {
-          '& .MuiOutlinedInput-notchedOutline, & .MuiFilledInput-underline, & .MuiInput-underline': {
-            borderColor: ({ options: { errorColor } }) => [
-              style.getColor(errorColor),
-              '!important',
-            ],
-          },
-          '&.MuiInput-underline, &.MuiFilledInput-underline': {
-            '&::before, &::after': {
-              borderColor: ({ options: { errorColor } }) => [
-                style.getColor(errorColor),
+          '&.Mui-focused::before, &.Mui-focused::after, &.Mui-focused:hover::before, &.Mui-focused:hover::after':
+            {
+              borderColor: ({ options: { borderFocusColor } }) => [
+                style.getColor(borderFocusColor),
                 '!important',
               ],
             },
-            '&:hover': {
+        },
+        '& .MuiInputBase-root.Mui-error, & .MuiInputBase-root.Mui-error:hover, & .MuiInputBase-root.Mui-error.Mui-focused, & .MuiInputBase-root.Mui-error.Mui-focused:hover':
+          {
+            '& .MuiOutlinedInput-notchedOutline, & .MuiFilledInput-underline, & .MuiInput-underline':
+              {
+                borderColor: ({ options: { errorColor } }) => [
+                  style.getColor(errorColor),
+                  '!important',
+                ],
+              },
+            '&.MuiInput-underline, &.MuiFilledInput-underline': {
               '&::before, &::after': {
                 borderColor: ({ options: { errorColor } }) => [
                   style.getColor(errorColor),
                   '!important',
                 ],
               },
-            },
-            '&.Mui-focused::before, &.Mui-focused::after, &.Mui-focused:hover::before, &.Mui-focused:hover::after': {
-              borderColor: ({ options: { errorColor } }) => [
-                style.getColor(errorColor),
-                '!important',
-              ],
+              '&:hover': {
+                '&::before, &::after': {
+                  borderColor: ({ options: { errorColor } }) => [
+                    style.getColor(errorColor),
+                    '!important',
+                  ],
+                },
+              },
+              '&.Mui-focused::before, &.Mui-focused::after, &.Mui-focused:hover::before, &.Mui-focused:hover::after':
+                {
+                  borderColor: ({ options: { errorColor } }) => [
+                    style.getColor(errorColor),
+                    '!important',
+                  ],
+                },
             },
           },
-        },
       },
     };
   },